Adjusted score prop type to Decimal

+++ src/model/orderedQuestion.js
@@ -31,7 +31,7 @@ export const orderedQuestion = {
         score: {
             required: false,
             multiple: false,
-            dataType: "integer",
+            dataType: "decimal",
         },
         comment: {
             required: false,
